make id long

+++ src/main/java/org/example/repository/AccidentRepository.java
@@ -4,9 +4,10 @@ import org.example.entity.Accident;
 import org.example.entity.Contract;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 @Repository
-public interface AccidentRepository extends JpaRepository<Accident, Integer> {
+public interface AccidentRepository extends JpaRepository<Accident, Long> {
     List<Accident> findAccidentByContract(Contract contract);
 }
